app/schemas/auth.py
from pydantic import BaseModel


class TokenSchema(BaseModel):
    access_token: str
    token_type: str


class TokenDataSchema(BaseModel):
    username: str
    hashed_password: str
    expires_at: str
