requirements.txt
fastapi[all]
sqlalchemy
python-dotenv
passlib
python-jose

bcrypt
alembic

pillow
torch==2.2.1+cu121
torchvision==0.17.1+cu121
scikit-learn
scikit-plot

scipy==1.11.4
schedule